flask-nutricional/requirements.txt
python-dotenv
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
codecov==2.1.13
colorama==0.4.6
coverage==7.6.8
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
fpdf==1.7.2
frozenlist==1.5.0
guardrails-ai
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
MarkupSafe==3.0.2
multidict==6.1.0
openai==0.28.0
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.3
